# backend/openapi_v3.yaml
openapi: 3.0.1
info:
  title: 团队心流引擎 (TeamFlow Engine) API
  description: 为《团队心流引擎》应用提供后端服务的RESTful API。该API支持团队管理、每日状态签到、数据可视化和AI分析等核心功能。
  version: 1.2.1
...
paths:
...
  /teams/:
    get:
      summary: 获取当前用户加入的团队列表
      deprecated: false
      description: ''
      operationId: listMyTeams
      tags:
        - Teams
      parameters: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
          headers: {}
      security: []
    post:
      summary: 创建一个新团队
      deprecated: false
      description: ''
      operationId: createTeam
      tags:
        - Teams
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
        required: true
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          headers: {}
      security: []
  /teams/{team_id}/:
    get:
      summary: 获取指定团队的详细信息
      deprecated: false
      description: ''
      operationId: retrieveTeam
      tags:
        - Teams
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
          headers: {}
      security: []
  /teams/{team_id}/invitations/:
    post:
      summary: 邀请成员加入团队
      deprecated: false
      description: ''
      operationId: inviteMember
      tags:
        - Teams
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 邀请已发送
          headers: {}
      security: []
  /teams/invitations/accept/:
    post:
      summary: 接受团队邀请
      deprecated: false
      description: ''
      operationId: acceptInvitation
      tags:
        - Teams
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - invite_code
              properties:
                invite_code:
                  type: string
      responses:
        '200':
          description: 成功加入团队
          headers: {}
      security: []
  /teams/{team_id}/checkins/:
    post:
      summary: 为指定团队提交今日签到
      deprecated: false
      description: ''
      operationId: createCheckin
      tags:
        - Check-ins
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinCreate'
      responses:
        '201':
          description: 签到成功
          headers: {}
      security: []
  /teams/{team_id}/checkins/today/:
    get:
      summary: 检查当前用户今天是否已签到
      deprecated: false
      description: ''
      operationId: checkTodayStatus
      tags:
        - Check-ins
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 返回今日签到状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_checked_in:
                    type: boolean
                required:
                  - has_checked_in
          headers: {}
      security: []
...
  /teams/{team_id}/highlights/:
    get:
      summary: 获取团队高光时刻列表
      deprecated: false
      description: ''
      operationId: listHighlights
      tags:
        - Recognition
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Highlight'
          headers: {}
      security: []
    post:
      summary: 发布一个新的高光时刻
      deprecated: false
      description: ''
      operationId: createHighlight
      tags:
        - Recognition
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '201':
          description: 发布成功
          headers: {}
      security: []
  /teams/{team_id}/flow-sessions/:
    post:
      summary: 提交一次心流仪式(专注)记录
      deprecated: false
      description: ''
      operationId: createFlowSession
      tags:
        - Flow Ritual
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowSessionCreate'
      responses:
        '201':
          description: 记录成功
          headers: {}
      security: []
...
  /teams/{team_id}/kudos/:
    post:
      summary: 发送一张Kudos能量卡
      deprecated: false
      description: ''
      operationId: sendKudos
      tags:
        - Recognition
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KudosCreate'
      responses:
        '201':
          description: 发送成功
          headers: {}
      security: []
...
  /teams/{team_id}/charter/:
    get:
      summary: 获取团队心流公约
      deprecated: false
      description: ''
      operationId: getCharter
      tags:
        - Culture & Growth
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charter'
          headers: {}
      security: []
    put:
      summary: 更新团队心流公约 (管理员)
      deprecated: false
      description: ''
      operationId: updateCharter
      tags:
        - Culture & Growth
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Charter'
      responses:
        '200':
          description: 更新成功
          headers: {}
      security: []
  /teams/{team_id}/skill-tree/:
    get:
      summary: 获取团队技能树数据
      deprecated: false
      description: ''
      operationId: getSkillTree
      tags:
        - Culture & Growth
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillTreeData'
          headers: {}
      security: []
...
components:
  schemas:
...
    FlowSessionCreate:
      type: object
      required:
        - start_time
        - duration_minutes
        - task_description
      properties:
        start_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        task_description:
          type: string
...
    KudosCreate:
      type: object
      required:
        - receiver_id
        - card_type
        - message
      properties:
        receiver_id:
          type: string
          format: uuid
        card_type:
          type: string
        message:
          type: string
    Charter:
      type: object
      properties:
        content:
          type: string
          description: Markdown格式的公约内容
        last_updated_by:
          $ref: '#/components/schemas/User'
        updated_at:
          type: string
          format: date-time
    SkillTreeData:
      type: object
      description: 适合ECharts Tree Graph的数据结构
      properties:
        name:
          type: string
          example: 团队技能
        children:
          type: array
          items:
            type: object
            properties: {}
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner:
          $ref: '#/components/schemas/User'
    TeamCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: 凤凰项目组
    TeamDetail:
      allOf:
        - $ref: '#/components/schemas/Team'
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
      type: object
...
    CheckinCreate:
      type: object
      required:
        - challenge_level
        - skill_level
      properties:
        challenge_level:
          type: number
          format: float
          description: 挑战感, -1 to 1
        skill_level:
          type: number
          format: float
          description: 能力感, -1 to 1
        achievement_text:
          type: string
          nullable: true
        obstacle_text:
          type: string
          nullable: true
...
    Highlight:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
        created_at:
          type: string
          format: date-time
        likes_count:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers: []
security: []